anyhow: clean up doc-comment references to deleted ErrorKind variants
Code-review follow-up to the strider-error → anyhow migration. No
runtime behaviour change; tests + clippy stay green.

- Rewrite stale `# Errors` blocks across strider, opt, reader, cfg
  that named deleted variants (ErrorKind::IrError, ErrorKind::Object,
  ErrorKind::IndirectResolutionDidNotConverge, …) into plain prose
  matching the actual anyhow message text.
- Update pattern::rewrite_rule's doc to describe the new
  free-function `skip` / `is_skip` API and downcast-via-anyhow
  custom-error path; drop refs to the deleted
  `ErrorKind::IrError` / `ErrorKind::RewriteClosure` /
  `Error::rewrite_closure`.
- Add a code comment on `GraphDotDumper::Error` explaining why the
  bound is `Debug + Display + Send + Sync + 'static` (and not
  `std::error::Error`): so impls can pick `type Error = anyhow::Error`,
  which deliberately does not implement `std::error::Error`.
- Drop two `u128::from(masked)` self-conversion clippy warnings in
  opt's constant_fold tests that crept in via a prior type widen.

+++ crates/opt/src/constant_fold/tests.rs
@@ -198,7 +198,7 @@ fn assert_add_with_const(
         matches!(
             *fg.graph.kind_of_output(o),
             // IntConst stores u128; masked is u64, widen for comparison.
-            NodeKind::IntConst(v) if ty.get_unsigned_int(v) == Some(u128::from(masked))
+            NodeKind::IntConst(v) if ty.get_unsigned_int(v) == Some(masked)
         )
     };
     let ok = (l == expected_base && const_on(r)) || (r == expected_base && const_on(l));
@@ -240,7 +240,7 @@ fn assert_sub_with_const(
     let const_on_rhs = matches!(
         *fg.graph.kind_of_output(r),
         // IntConst stores u128; masked is u64, widen for comparison.
-        NodeKind::IntConst(v) if ty.get_unsigned_int(v) == Some(u128::from(masked))
+        NodeKind::IntConst(v) if ty.get_unsigned_int(v) == Some(masked)
     );
     assert!(
         l == expected_base && const_on_rhs,
